move component behavior to container
